+++ src/app/page.tsx
@@ -1,7 +1,7 @@
 import Image from 'next/image'
 import { Inter } from "next/font/google";
 import styles from './page.module.css'
-import User from "./user/page";
+import User from "./users/[id]/page";
 
 const inter = Inter({ subsets: ['latin'] })
